+++ transposcope_pkg/fasta_handler.py
