+++ transposcope_pkg/parsers/tipseqhunter_parser.py
