+++ transposcope_pkg/web/js/scatter_plotter.js
